Moving new submission actions into thier own controller
