+++ app/views/wizard_new_submission/new_submission.html.erb
@@ -36,7 +36,7 @@
   <%= render '/works/form_errors' %>
 
   <h1>New Submission</h1>
-  <%= render "wizard_progress", wizard_step: 0 %>
+  <%= render "works_wizard/wizard_progress", wizard_step: 0 %>
 
   <p>
     You are starting a new submission.
